Refactoring Patient and Chair files
