+++ db/crud.py
@@ -1,6 +1,6 @@
 from sqlalchemy.orm import Session
 import db.schemas as schemas, db.models as models
-from fastapi import HTTPException, status, Depends
+from fastapi import HTTPException, status
 from fastapi_jwt_auth import AuthJWT
 from fastapi.encoders import jsonable_encoder
 from passlib.context import CryptContext
@@ -63,235 +63,6 @@ def generate_tokens(id: int, authorize: AuthJWT):
         "message": "Success",
     }
     return jsonable_encoder(response)
-
-
-def chair_signup(db: Session, chair: schemas.ChairRegistration):
-    """
-    We use this function to store the new chair in the database
-    We first check if the id that send with the data is already stored in the database or not
-    if it stored we raise an exception if not we stored the chair in our database
-
-    Args:
-        db (Session): The database session that we will use to store the data
-        chair (schemas.ChairRegistration): The data that we want to store in the database (chair_id, password)
-
-    Raises:
-        HTTPException: If the chair_id already exist in the database we raise exception with status code 400
-
-    Returns:
-        Dict : we return a message tell us that we store the chair data successfully
-    """
-
-    db_chair = db.query(models.Chair).filter(chair.chair_id == models.Chair.id).first()
-
-    if db_chair is not None:
-        raise HTTPException(
-            status_code=status.HTTP_400_BAD_REQUEST,
-            detail="This Chair ID is already exist",
-        )
-
-    new_chair = models.Chair(
-        id=chair.chair_id, password=create_hashed_password(chair.password)
-    )
-
-    db.add(new_chair)
-    db.commit()
-    db.refresh(new_chair)
-
-    return {"message": "Chair register successfully"}
-
-
-def get_chair(db: Session, chair: schemas.ChairRegistration):
-    """
-    This function validate the data that have been sent to login to specific chair
-    First we check if the chair_id is exist in the database then verify and compare
-    the password that send with the request
-    If everything go well then We return a message tell us the login done successfully
-    if not then return None which means noting found
-
-    Args:
-        db (Session): The database session that we will use to validate the data
-        chair (schemas.ChairRegistration): The data that we want to validate with stored data (chair_id, password)
-
-    Returns:
-        Dict : We return a message tell us the login done successfully
-        None : Return None when no chair exist in the database with this ID and password
-    """
-
-    current_chair = (
-        db.query(models.Chair).filter(models.Chair.id == chair.chair_id).first()
-    )
-
-    if current_chair and verify_password(chair.password, current_chair.password):
-        return {"message": "Chair login successfully"}
-    return None
-
-
-def chair_login(db: Session, chair: schemas.ChairRegistration):
-    """
-    We use this function to login or access the chair with a specific id
-
-    Args:
-        db (Session): The database session that we will use to validate the data
-        chair (schemas.ChairRegistration): The data that we want to validate with stored data (chair_id, password)
-
-    Raises:
-        HTTPException: If the chair_id not exist in the database we raise an exception with status code 404
-
-    Returns:
-        Dict : We return a message tell us the login done successfully
-    """
-
-    current_chair = get_chair(db=db, chair=chair)
-    if current_chair is None:
-        raise HTTPException(
-            status_code=status.HTTP_404_NOT_FOUND, detail="Chair ID or Password Invalid"
-        )
-    return current_chair
-
-
-def store_chair_data(db: Session, data: schemas.ReadChairData):
-    """
-    This fucntion used to store the data coming from sensors but only when the chair is existing in the database
-    if the chair not found we raise an exception, otherwise we store the data
-
-    Args:
-        db (Session): The database session that we will use to validate the data
-        data (schemas.ReadChairData): The sensors' data that we want to store in the database
-
-    Raises:
-        HTTPException: We check if the chair is exist in the database, if not we raise an exception with code 404
-
-    Returns:
-        Dict : we return a message tell us that the data stored successfully
-    """
-
-    db_chair = db.query(models.Chair).filter(models.Chair.id == data.chair_id).first()
-
-    if db_chair is None:
-        raise HTTPException(
-            status_code=status.HTTP_404_NOT_FOUND, detail="No Chair stored with this ID"
-        )
-
-    new_data = models.SensorData(**data.dict())
-
-    chair = db.query(models.Chair).filter(data.chair_id == models.Chair.id).first()
-
-    new_data.chair = chair
-
-    db.add(new_data)
-    db.commit()
-    db.refresh(new_data)
-
-    return {"message": "Data has been stored successfully"}
-
-
-def get_chair_data(chair_id: int, db: Session):
-    """
-    This function used to get the latest data of specific chair that have been stored using the chair_id
-    that send with the request to the route and return the data if the chair_id exist
-
-    Args:
-        chair_id (int): The id of the chair we want to connect to access his data
-        db (Session): The database session that we will use to validate the data
-
-    Raises:
-        HTTPException: If the ID not found we raise an exception with status code 404
-
-    Returns:
-        Dict: We return the data of the sonsors connected to the chair that have id = chair_id
-    """
-    sensor_data = (
-        db.query(models.SensorData)
-        .filter(models.SensorData.chair_id == chair_id)
-        .order_by(models.SensorData.id.desc())
-        .first()
-    )
-
-    if sensor_data is None:
-        raise HTTPException(
-            status_code=status.HTTP_404_NOT_FOUND, detail="Data not found"
-        )
-
-    return schemas.GetChairData.from_orm(sensor_data)
-
-
-def add_new_patient(db: Session, patient: schemas.PatientDataRegister):
-    """
-    We use this function to store the patient's data in the database
-    with the chair_id they use
-
-    Args:
-        db (Session): The database session that we will use to validate the data
-        patient (schemas.PatientData): The patient's data that we want to store in the database
-        chair_id (int): The chair_id of the chair that the patient use and stored in the token
-
-    Returns:
-        Dict: we return a Dict message tell us we stored patient's data successfully
-    """
-
-    chair_info = {"chair_id": patient.chair_id, "password": patient.password}
-
-    login_chair_schema = schemas.ChairRegistration(**chair_info)
-
-    db_chair = chair_login(db=db, chair=login_chair_schema)
-
-    new_patient = models.Patient(
-        first_name=patient.first_name,
-        last_name=patient.last_name,
-        gender=patient.gender,
-        age=int(patient.age),
-        chair_id=patient.chair_id,
-    )
-
-    chair = db.query(models.Chair).filter(models.Chair.id == patient.chair_id).first()
-
-    new_patient.chair = chair
-
-    db.add(new_patient)
-    db.commit()
-    db.refresh(new_patient)
-
-    return new_patient
-
-
-def patient_info(chair_id: int, db: Session):
-    db_patient = (
-        db.query(models.Patient).filter(chair_id == models.Patient.chair_id).first()
-    )
-
-    if db_patient is None:
-        raise HTTPException(
-            status_code=status.HTTP_404_NOT_FOUND,
-            detail="No Patient currently use this chair",
-        )
-
-    return db_patient
-
-
-def update_patient_chair(
-    current_chair_id: int, new_chair: schemas.ChairRegistration, db: Session
-):
-    db_patient = (
-        db.query(models.Patient)
-        .filter(models.Patient.chair_id == current_chair_id)
-        .first()
-    )
-
-    login_to_new_chair = chair_login(db=db, chair=new_chair)
-
-    db_patient.chair_id = new_chair.chair_id
-
-    db_chair = (
-        db.query(models.Chair).filter(models.Chair.id == new_chair.chair_id).first()
-    )
-
-    db_patient.chair = db_chair
-
-    db.commit()
-    db.refresh(db_patient)
-
-    return db_patient
 
 
 def signup_caregiver(
